Move Ast-related MakeType to Ast.Import
