+++ packages/decoder/lib/decoders.ts
@@ -162,10 +162,7 @@ export class WireDecoder {
       //first, add the contract itself
       const contractNode = this.contractNodes[id];
       references[id] = contractNode;
-      types[id] = Format.Utils.MakeType.definitionToStoredType(
-        contractNode,
-        compiler
-      );
+      types[id] = Ast.Import.definitionToStoredType(contractNode, compiler);
       //now, add its struct and enum definitions
       for (const node of contractNode.nodes) {
         if (
@@ -175,7 +172,7 @@ export class WireDecoder {
           references[node.id] = node;
           //HACK even though we don't have all the references, we only need one:
           //the reference to the contract itself, which we just added, so we're good
-          types[node.id] = Format.Utils.MakeType.definitionToStoredType(
+          types[node.id] = Ast.Import.definitionToStoredType(
             node,
             compiler,
             references
